No warnings at all. Doesn't die with VIDEO_ZBUF on.

+++ trunk/phantom/vm/syscall_tty.c
@@ -132,7 +132,7 @@ static int debug_18(struct pvm_object me , struct data_area_4_thread *tc )
 
     //pvm_object_print( o );
     printf("\n\nobj dump: ");
-    dumpo(o.data);
+    dumpo((int)(o.data));
     printf("\n\n");
 
     SYS_FREE_O(o);
